Added UserEntity database test

+++ app/src/main/java/com/quoders/app/codewarschallenge/data/local/entities/UserEntity.kt
@@ -17,5 +17,5 @@ class UserEntity {
     var clan: String? = null
 
     @ColumnInfo(name = "honor")
-    var honor: Int = 0
+    var honor: Long = 0
 }
